Restructure helpers a little bit
