--- module/dice/_module.mjs
@@ -1,1 +1,0 @@
-export { default as Roll4e } from "./Roll.mjs";
