+++ module/rolls/_module.mjs
@@ -1,0 +1,3 @@
+export { default as Roll4e } from "./roll.mjs";
+export { default as MultiAttackRoll } from "./multi-attack-roll.mjs";
+export { default as RollWithOriginalExpression } from "./roll-with-expression.mjs";
